add tokenhandlers and check for admin role

+++ server/controllers/productsController.js
@@ -1,6 +1,6 @@
 const ErrorResponse = require("../utils/errorRepsonse");
 const Product = require("../models/ProductModel");
-const { findByIdAndUpdate } = require("../models/ProductModel");
+const User = require("../models/UserModel")
 
 async function getAllProducts(req, res, next) {
   try {
@@ -37,9 +37,14 @@ async function updateProduct(req, res, next) {
   try {
     const productId = req.params.id;
     const productToUpdate = await Product.findById(productId);
+    const user = await User.findById(req.userId);
+
+    if (user.role !== "admin") {
+      return next(new ErrorResponse("Not authorized", 401))
+    }
 
     if (!productToUpdate) {
-      return next(new ErrorResponse("Product not found", 401))
+      return next(new ErrorResponse("Product not found", 404))
     }
 
     const updateProduct = await Product.findByIdAndUpdate(productId, req.body, {
@@ -60,11 +65,15 @@ async function updateProduct(req, res, next) {
 async function deleteProduct (req, res, next) {
   try {
     const productId = req.params.id;
-  
     const productToDelete = await Product.findById(productId);
+    const user = await User.findById(req.userId);
+
+    if (user.role !== "admin") {
+      return next(new ErrorResponse("Not authorized", 401))
+    }
   
     if (!productToDelete) {
-      return next(new ErrorResponse("Product not found", 401))
+      return next(new ErrorResponse("Product not found", 404))
     }
   
     await Product.findByIdAndDelete(productId);
@@ -78,9 +87,30 @@ async function deleteProduct (req, res, next) {
   }
 }
 
+
+async function createNewProduct (req, res, next) {
+  try {
+    const product = req.body;
+    const user = await User.findById(req.userId);
+
+    if (user.role !== "admin") {
+      return next(new ErrorResponse("Not authorized", 401))
+    }
+
+    await Product.create(product);
+    res.status(200).json({
+      success: true,
+      product: product,
+    });
+  } catch (err) {
+    next(err);
+  }
+}
+
 module.exports = {
   getAllProducts,
   getSingleProduct,
   updateProduct,
-  deleteProduct
+  deleteProduct,
+  createNewProduct
 };
